requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
pandas==2.1.4
python-dotenv==1.0.0
scikit_learn==1.3.2
tensorflow==2.16.0
